+++ doc/General/ChangeLog.md
@@ -6,6 +6,7 @@ Mirror uses semantic versioning, and the versions shown here are those that were
 
 ## Version 11.x.x - In Progress
 - Added: SyncVar hooks can be virtual now, and overriden in a derived class.
+- Added: Virtual OnRoomStopServer to NetworkRoomManager and Script Template
 - Fixed: Setting breakpoints in an IDE for Command's and Rpc's work correctly now.
 - Fixed: NetworkServer's calls to SendToObservers now reports correct channel to Mirror Profiler.
 - Fixed: NetworkRoomManager's `roomPlayerPrefab` is now `protected` so it can be accessed in derived classes.